+++ aftiktuna-rust/src/serialization.rs
@@ -178,6 +178,7 @@ pub mod world {
         core::CrewMember, CrewMember;
         core::Hostile, Hostile;
         core::Wandering, Wandering;
+        core::ObservationTarget, ObservationTarget;
         ai::Intention, Intention;
         core::Waiting, Waiting;
         core::RepeatingAction, RepeatingAction;